Finally fixed batch rendering pipeline. Fixed accidentally disabled error-logger. Changed batch shader to support switching between texture rendering and using vertex-shader (one of the two only) without 'if'. Code still contains numerous debugging and test lines - to be removed in the next commit.

+++ OpenGLTester/source/BaseClasses/Public/IndexBuffer.h
@@ -30,7 +30,7 @@ private:
     
     GLuint m_rendererId;
     size_t m_count;
-    size_t m_adjustment = 0;
+    GLuint m_adjustment = 0;
 
-    std::vector<size_t> m_vertexIndicesPool; 
+    std::vector<GLuint> m_vertexIndicesPool; 
 };
